Add display of file record
