+++ examples/eSealing/Jenkinsfile
@@ -4,7 +4,7 @@
 pipeline {
     agent any
     environment {
-        image_repo = "concordium/dapp-wccd"
+        image_repo = "concordium/dapp-eSealing"
         image_name = "${image_repo}:${image_tag}"
     }
     stages {
@@ -26,7 +26,7 @@ pipeline {
                       --label build_image="${build_image}" \
                       --tag="${image_name}" \
                       --pull \
-                      ./examples/wCCD
+                      ./examples/eSealing
                     docker push "${image_name}"
                 '''.stripIndent()
             }
