+++ examples/eSealing/README.md
@@ -1,16 +1,20 @@
-# wCCD web application
+# eSealing web application
 
-The example project included in this repository serves as a working example of how to integrate with smart contracts on the Concordium blockchain. This web app interacts with the wCCD protocol on testnet. The connected account can invoke the wrap/unwrap functions to observe the wCCD balance change of its account.
+The example project included in this repository serves as a working example of how to integrate with smart contracts on the Concordium blockchain. This web app supports the following two flows with the browser wallet (or wallet connect):
+
+-   Upload a file from the computer => register its file hash in the smart contract
+-   Upload a file from the computer => retrieve the time-stamp and witness (sender_account) that registered the file hash
 
 ## Prerequisites
 
--   Browser wallet extension must be installed in google chrome and configured with testnet JSON-RPC, in order to view smart contract details or submit transactions.
+-   Browser wallet extension must be installed in google chrome and configured with testnet JSON-RPC or a mobile wallet needs to be set up that supports wallet connect in order to view smart contract details or submit transactions.
 
 ## Installing
 
 -   Run `yarn` in package root.
+-   Run `yarn build:all` in package root.
 
-## Running the wCCD example
+## Running the eSealing example
 
 -   Run `yarn build` in a terminal in this folder.
 -   Run `yarn start`.
@@ -27,25 +31,25 @@ To have hot-reload (useful for development), do the following instead:
 To build the docker image run the following:
 
 ```
-docker build -t wccd_front_end:$PROJECT_VERSION .
+docker build -t eSealing_front_end:$PROJECT_VERSION .
 ```
 
 e.g.
 
 ```
-docker build -t wccd_front_end:3.0.0 .
+docker build -t eSealing_front_end:3.0.0 .
 ```
 
 To run the docker image run the following:
 
 ```
-docker run -it -d -p 8080:80 --name web wccd_front_end:$PROJECT_VERSION
+docker run -it -d -p 8080:80 --name web eSealing_front_end:$PROJECT_VERSION
 ```
 
 e.g.
 
 ```
-docker run -it -d -p 8080:80 --name web wccd_front_end:3.0.0
+docker run -it -d -p 8080:80 --name web eSealing_front_end:3.0.0
 ```
 
 Open http://127.0.0.1:8080 in your browser.
