+++ examples/eSealing/src/Root.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import { SignClientTypes } from '@walletconnect/types';
-import { version } from '../package.json';
 
-import WCCD from './wCCD';
+import SEALING from './eSealing';
 import { WithWalletConnection } from './wallet/WithWalletConnection';
 import { TESTNET, WALLET_CONNECT_PROJECT_ID } from './constants';
 import { Network } from './wallet/WalletConnection';
@@ -11,8 +10,8 @@ const network: Network = TESTNET;
 const walletConnectOpts: SignClientTypes.Options = {
     projectId: WALLET_CONNECT_PROJECT_ID,
     metadata: {
-        name: 'wCCD',
-        description: 'Example dApp for the wCCD token.',
+        name: 'eSealing',
+        description: 'Example dApp for the eSealing a file.',
         url: '#',
         icons: ['https://walletconnect.com/walletconnect-logo.png'],
     },
@@ -26,21 +25,10 @@ export default function Root() {
         <div>
             <script src="../../node_modules/@concordium/web-sdk/lib/concordium.min.js" />
             <script src="../../packages/browser-wallet-api-helpers/lib/concordiumHelpers.min.js" />
-            <main className="wccd">
+            <main className="eSealing">
                 <WithWalletConnection network={network} walletConnectOpts={walletConnectOpts}>
-                    {(props) => <WCCD {...props} />}
+                    {(props) => <SEALING {...props} />}
                 </WithWalletConnection>
-                <div>
-                    Version: {version} |{' '}
-                    <a
-                        style={{ color: 'white' }}
-                        href="https://developer.concordium.software/en/mainnet/smart-contracts/tutorials/wCCD/index.html"
-                        target="_blank"
-                        rel="noreferrer"
-                    >
-                        Learn how to make a wrapper like this
-                    </a>
-                </div>
             </main>
         </div>
     );
