+++ examples/eSealing/src/index.css
@@ -9,7 +9,7 @@ body {
     text-align: left;
 }
 
-.wccd {
+.eSealing {
     height: 100vh;
     width: 100%;
     display: flex;
@@ -25,8 +25,17 @@ body {
     letter-spacing: 0;
 }
 
+.centerLargeText {
+    font-size: 26px;
+    font-weight: 300;
+    line-height: 21px;
+    letter-spacing: 0;
+    text-align: center;
+}
+
 .loadingText {
     color: #979797;
+    overflow-wrap: anywhere;
 }
 
 .text {
@@ -41,7 +50,6 @@ button.buttonInvisible {
 }
 
 .header {
-    margin-top: 0;
     color: white;
 }
