+++ examples/eSealing/src/index.html
@@ -2,7 +2,7 @@
 <html>
     <style>
         body {
-            background-image: url(./Concordium_background.svg);
+            background-image: url(./Concordium_background.jpg);
             background-repeat: no-repeat;
             background-attachment: fixed;
             background-size: cover;
